+++ example/pathlib_example.dart
@@ -2,9 +2,20 @@
 import 'package:pathlib/pathlib.dart';
 
 main() {
-  var p = Path("lib");
-  p += Path("src/pathlib_base.dart");
-  print(p.abs);                                    // => /home/jwinnie/code/dart/pathlib/lib/src/pathlib_base.dart
-  print(Path("/etc") == Path("/etc"));             // => true
-  print(Path("/etc/../") == Path("/etc/passwd"));  // => false
+
+  // Easily create paths
+  Path("/etc/passwd");
+  Path.cwd(); // current directory
+  Path.root(); // root directory
+
+  // Easily work with paths
+  Path("/etc") + Path("passwd");
+
+  // Easily convert relative to absolute paths
+  Path("lib/src/").abs;
+
+  // Easily iterate paths
+  for (var i in Path("/etc").list()) {
+    print(i);
+  }
 }