+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+- Added `iter()` method to list files in a directory
+- Added `hashCode` override
+- Documentation improvements
+
+## 1.0.1
+- Documentation update
+
 ## 1.0.0
 Initial version!
 - Created Path class